apps/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
argon2-cffi>=23.1.0
python-multipart>=0.0.6
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
httpx>=0.26.0

# AI/ML
groq>=0.4.0
llama-index>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
chromadb>=0.4.22
langchain>=0.1.0
langgraph>=0.0.20
tiktoken>=0.5.2

# AST Parsing
tree-sitter>=0.20.4
tree-sitter-python>=0.20.1
tree-sitter-javascript>=0.20.1
tree-sitter-typescript>=0.20.1

# Git Operations
GitPython>=3.1.40

# Graph Analysis & Database
networkx>=3.2
neo4j>=5.15.0

# Code Analysis
radon>=6.0.1

# Utils
python-dotenv>=1.0.0
structlog>=24.1.0
